extract remote-access module
