+++ packages/unraid-api-plugin-connect/src/module/connect.module.ts
@@ -9,7 +9,7 @@ import { ConnectConfigService } from '../config/connect.config.service.js';
 import { ConnectSettingsResolver } from '../resolver/connect-settings.resolver.js';
 import { ConnectResolver } from '../resolver/connect.resolver.js';
 import { ConnectSettingsService } from '../service/connect-settings.service.js';
-import { RemoteAccessModule } from './remote-access.module.js';
+import { RemoteAccessModule } from '../remote-access/remote-access.module.js';
 
 @Module({
     imports: [RemoteAccessModule, ConfigModule, UserSettingsModule],
