+++ packages/unraid-api-plugin-connect/src/remote-access/remote-access.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 
 import { WanAccessEventHandler } from '../event-handler/wan-access.events.js';
-import { DynamicRemoteAccessService } from '../service/dynamic-remote-access.service.js';
-import { StaticRemoteAccessService } from '../service/static-remote-access.service.js';
-import { UpnpRemoteAccessService } from '../service/upnp-remote-access.service.js';
-import { SystemModule } from './system.module.js';
+import { DynamicRemoteAccessService } from './dynamic-remote-access.service.js';
+import { StaticRemoteAccessService } from './static-remote-access.service.js';
+import { UpnpRemoteAccessService } from './upnp-remote-access.service.js';
+import { SystemModule } from '../module/system.module.js';
 
 @Module({
     imports: [SystemModule],
